requirements.txt
flask
flask-cors
python-dotenv
waitress
requests
openai
